backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
Werkzeug==2.3.7
PyJWT==2.8.0
pdfkit==1.0.0
Pillow==10.0.0
python-dotenv==1.0.0
email-validator==2.0.0
gunicorn==21.2.0
SQLAlchemy==2.0.21
pytest==7.4.1
pytest-flask==1.2.0
